Move pacbio library request from tubes vuex, to libraries
