+++ src/components/PacbioLibrariesList.vue
@@ -12,7 +12,7 @@ import Helper from '@/mixins/Helper'
 import TableHelper from '@/mixins/TableHelper'
 import { createNamespacedHelpers } from 'vuex'
 // TODO: move to pacbio libraries
-const { mapActions, mapGetters } = createNamespacedHelpers('traction/pacbio/tubes')
+const { mapActions, mapGetters } = createNamespacedHelpers('traction/pacbio/libraries')
 import Tube from '@/components/Tube'
 
 export default {
