+++ src/store/traction/pacbio/libraries/getters.js
@@ -1,0 +1,6 @@
+const getters = {
+  libraryRequest:  (state, getters, rootState) => rootState.api.traction.pacbio.libraries,
+  libraries: state => state.libraries
+}
+
+export default getters
