+++ src/store/traction/pacbio/libraries/index.js
@@ -1,0 +1,14 @@
+import state from './state'
+import getters from './getters'
+import mutations from './mutations'
+import actions from './actions'
+
+const libraries = {
+  namespaced: true,
+  state,
+  getters,
+  mutations,
+  actions
+}
+
+export default libraries
