+++ src/store/traction/pacbio/libraries/state.js
@@ -1,5 +1,5 @@
 const state = {
-  tractionTubes: []
+  libraries: []
 }
 
 export default state
