+++ src/store/traction/pacbio/tubes/mutations.js
@@ -1,15 +1,6 @@
 const mutations = {
   setTubes (state, tubes) {
     state.tractionTubes = tubes
-  },
-  setLibraries(state, libraries) {
-    state.libraries = libraries
-  },
-  updateLibrary(state, library) {
-    let index = state.libraries.findIndex(lib => lib.id == library.id)
-    let libraries = state.libraries
-    libraries.splice(index, 1, library)
-    state.libraries = libraries
   }
 }
