+++ src/views/pacbio/PacbioReception.vue
@@ -65,6 +65,7 @@ export default {
         this.showAlert(error.message, 'danger')
       }
     },
+    // TODO: move to requests
     ...mapActions('traction/pacbio/tubes', [
       'exportSampleExtractionTubesIntoTraction'
     ]),
